pyproject.toml
[tool.poetry]
name = "netiflix-data"
version = "0.1.0"
description = "Pipeline de dados profissional"
authors = ["Paulo Filho <paulocarlosfilho@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pyarrow = ">=15.0.0"
setuptools = "^70.0.0"
luigi = "^3.5.0"
pandas = "^2.2.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
scikit-learn = "^1.3.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
numpy = "<2.0.0"
prometheus-client = "^0.24.1"
great-expectations = "<1.0.0"
openpyxl = "^3.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"